Add protected '/dashboard/' route and component
The dashboard page is a throwaway page to quickly browse through
the redux store and api state. Please ignore non-standard practices
(like not having a select component in its own file, having a lot of
code for one component, etc.) Those good practices will be used once
the development of the actual UI begings.
The way this page works is that when the user selects an organization,
the reporting periods, the organization units, and the EF dataset
of the selected organization are automatically fetched. Also, as an
example of fetching manually on demand, there is a button at the end of
the page to fetch the user organizations, using the useLazyQuery from
RTK Query.